reactor: lower the default stall threshold to 200ms
Historically, the stall detector threshold was set to 2s because when
introduced, we were afraid that lower values would spam the logs too
much.  We set it to 2s because "it's almost impossible that it will
hit", and any significant stall events should be very rare.

What we saw over the past two years, is that we were as wrong as Bill
Gates when he said that "640k ought to be enough for everybody". Stalls
are common, happen from the craziest of reasons, and can come from
anywhere in the platform.

It has evolved to be one of the main tools in debugging latency issues
that show up in user environments. The default of 2s however makes us
blind to occurrences that are by all means still bad - in the order of
hundreds of milliseconds unless we go and explicitly search for them.

For events that are hard to reproduce, this may be too late.

In the context of Scylla, I have been advocating for us to lower the
stall threshold for a long while now. Because this is a Scylla-specific
ask, I considered adding machinery to allow the application to override
seastar defaults like we do with other knobs like the task quota.

However upon consideration, I will argue that any seastar application
would benefit from the lower default and increased visibility on
blocking events.

One of the main sources of false positives, external applications
stealing time from the reactor, was recently fixed by Avi. So there's
no better time than now to lower the threshold and usher a new era
of lower predictable latencies.

We've been fixing stalls all over Seastar and Scylla for years. If there
are any left, let's fix them. If the log activity increases, let us look
them in the eyes and say, with conviction: "you don't belong here". But
let's start by being aware of them.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20191122154810.24082-1-glauber@scylladb.com>