reactor: add abandoned failed futures counter
Count the instances of futures containing an unhandled exception being
destroyed. We are already logging these instances as they most likely
point to bugs in the application. However logging doesn't provide enough
visibility as one has to be actively monitor the logs to observe this
happening. Provide a counter to allow better observability.