testing: propagate failures to the boost test framework gracefully
Currently the way the seastar framework can propagate any failure to the
boost test framework and ultimately back to the user running the test
executable is broken. We just call `exit()` in the thread hosting the
`app::run_deprecated()` function, if the latter has returned with a
non-zero exit code. This can lead to segfault depending on the phase the
test application is in.

Instead of communicating failures by yanking the ground from below the
feet of the entire application, `test_runner` now returns the success of
its run phases from `start()` and `finalize()` respectively.

`test_runner::start()` now waits for the seastar app to initialize and
returns whether this was successful or not. This status code is then
propagated back to the boost test framework by returning it from
`unit_unit_test_suite()` instead of unconditionally returning `true`.
This means that if the seastar app failed to initialize we abort the
test process in the initialization phase, before we attempt to run any
of the test cases.

`test_runner::finalize()` now returns an exit code, the one provided by
the `seastar::app::run()`. This exit code is then considered by
`entry_point()` when returning back to the caller (`main()`). If both
`boost::unit_test::unit_test_main()` and `test_runner::finalize()`
returns a non-zero error code, the boost one is preferred.