future: Move make_ready out of line
It is not clear why this was marked always_inline. That was done in
8812504e4, but that commit talks about futures and avoiding moves,
which I don't think applies in here.

Given what make_ready does, the one thing that seems possible to
optimize by inlining is the "if (_task)" check, but that is not at
least obviously known at the callers.

This reduces the size of .text in a release build of scylla from
42575908 to 41968068 bytes (1.4%).

The result of "perf_simple_query-patch --duration 16 --smp 1 -m4G"
improves just a bit, probably in the noise:

master:
median 147162.53
median absolute deviation: 193.05
maximum: 147853.55
minimum: 146723.50

patch:
median 147648.17
median absolute deviation: 158.74
maximum: 147889.10
minimum: 146762.24

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20191203201328.4803-1-espindola@scylladb.com>