reactor: fix iocb pool underflow due to unaccounted aio fsync
The iocb pool is sized for the worst case of all I/O
queues having 128 outstanding requests each. But fsync also consumes
iocbs from the pool, without any reservation having been made for
them. Usually this doesn't matter because those fsyncs complete
quickly, and because the pool is hugely oversized for the common
case. But in the uncommon case (ENOSPC) things can clog and we can
underflow the pool.

Fix by introducing a semaphore for guarding the pool. It should
be very rare that we have to wait on the semaphore, since usually
the io_queue mechanism will limit requests to well below the
pool capacity. It should only matter when the disk stalls and
gets filled with fsync requests.

Ref scylladb/scylla#5443.