future: Add a set_exception that takes a future_state_base
This takes advantage of the future_state_base::any move constructor
optimization that avoids calling the exception_ptr destructor on moved
values.

This reduces the number of ~exception_ptr calls instantiated in
future::then from 2 to 1.

The last call is from the future_state_base destructor. It might make
sense to move it out of line, but that requires more careful
benchmarking.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>