http: Change exception formating to the generic seastar one
It is beneficial to get the full context that an exception
can provide. When http translates an exception into an
http responce json, it ignored the case of nested exception
and always printed the last exception.
This patch makes use of seastars ability to print out exception
to the log along with types and nested exceptions information.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20191219151307.21441-2-eliransin@scylladb.com>