task: stop wrapping tasks with unique_ptr
Since the very beginning, we wrapped tasks and continuations
with unique_ptr for exception safety. It wasn't safe, however -
once a task is created, it must be executed and cannot be just
destroyed and freed, since anything that depends on the task
will not be run.

This became even more pronounced with 4d811de1463ba5059 ("task:
resumable tasks"), as tasks are now responsible for their own
reclamation (the "dispose" part of task::run_and_dispose()).

To make it clear that one cannot rely on unique_ptr<task> being
safe, replace unique_ptr<task> with a plain task* (and make_unique
with naked new calls).

To mark the fact that the functions in which this happens are not
exeception safe, annotate them as noexcept.

Some notes:
 - std::move()ing a pointer will not clear it, so some call sites
   have changed to use std::exchange()
 - repeat(), do_until(), and the like were using unique_ptr(this) to
   destroy the task in run_and_dispose() unless moved from. This
   was changed to plain "delete this".

Tests: unit(dev, debug)
Message-Id: <20191221210825.3179484-1-avi@scylladb.com>