reactor: Relax signals::handle_signal dependency on engine
In the next patch I'm going to call the signals::handle_signal helper
at an early stage, when the engine()._backend will not yet be ready.
For this I need to make the signal_handler constructor not mess with the
pointer in question.

To acheive this the most convenient way is to make the signal
handler registration happen in signals::handle_signal which also
lets factoring out epoll and aio backends' respective calls.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>