reactor: Setup timer sighandler earlier
Earler means -- backend constructor. To do this, we need to shuffle few
more bits, in particular the _backend pointer initialization is done in
contructor body to let _signals default constructor take place.

Also, the explicit blocking of timer signal in reactor constructor is
moved to aio backend constructor, as it's only needed there. Strictly
speaking this blocking was not required at all, even before this patch,
as all signals are blocked by smp::configure before reactors are created.
But let's make reactor not depend on this pre-requirement.

The order of pollers no longer matters, so the mentioned comment can be
removed. Also, the _timer_enabled is no longer needed.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>