json: add move assignment to json_return_type
The json_return_type type strangely has only a move constructor, but not
move assignment. This makes it "almost" usable in complex code, until
suddenly you get a 1000 line error message that tries to tell you, in
a very obfuscated way, that some code tried to move a json_return_type
using an assignment - but that wasn't supported.

So let's add the missing move assignment.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20200120163554.28930-1-nyh@scylladb.com>