core/memory.hh: document set_heap_profiling_enabled()
And make it discoverable in the generated documentation. For this it has
to be moved out from `/cond internal`.

The heap profiling data is not very consumable or useful in raw form so
provide links to an example way of processing it and presenting it in a
human-friendly way. This example is from the downstream application
Scylla.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20200123083454.195320-2-bdenes@scylladb.com>