smp: smp_message_queue::submit_item(): handle failure to acquire semaphore
Currently failures happening during waiting for the smp service groups's
semaphore are ignored. Not really a problem currently, as no timeout is
used when waiting for the semaphore, although `std::bad_alloc` is always
on the table.

To make this is safer, and to enable adding a timeout to the semaphore,
forward any failure to acquire the semaphore to the work item, and thus
to the caller of `smp::submit_to()`.