smp: make the smp service group semaphore a named semaphore
In the next patch we will add a timeout to said semaphore, which means
that error messages generated by it will appear in the application logs.
To help make these error messages more useful, use the named exception
factory and generates names that contain the smp service group, the
initiator (client) and receiver (server) shards.
Another forward looking change contained in this patch is defining and
using an smp specific timeout clock.