io_desc: split the io_desc into a generic, and promised version
Some uses of the io_desc can be expected to be so basic that
they won't even need to set promises. Extract the most basic
interface, and rename it to kernel_completion. That will be
used for anything that returns from the kernel.

The uses that have to set promises need to specialize it
themselves.

Signed-off-by: Glauber Costa <glauber@scylladb.com>