reactor: rework pollable_fd_state so it no longer holds promises
the pollable_fd_state holds two promises: pollin and pollout,
that are re-set every time we will do a poll operation.

Replace that with a promise_io_desc. That has a promise inside it,
and has the advantage that we will be able to call its completion
event from the return of io_getevents the same way we do for I/O.

A logical next step is to unify the handling of I/O and poll iocbs.

Signed-off-by: Glauber Costa <glauber@scylladb.com>