task: optimize destruction by making destructor non-virtual
task destruction is the responsibility of task::run_and_dispose() (which
usually just calls "delete this" at the appropriate time). Since it
knows the concrete type of the task, it does not need a virtal destructor.

Note that the compiler should already optimize this when it sees a
"delete this" in a final class as the concrete type is known, but making
the destructor non-virtual also reduces the vtable footprint. It is also
made protected to reduce the risk of destruction through a base class
pointer.

All classes implementing run_and_dispose() are now marked final to
reduce the chance of misuse.

Test: unit(dev)
Message-Id: <20200202104952.191061-1-avi@scylladb.com>