smp: reduce continuations and allocations in smp::submit_to()
with_scheduling_group() allocates and runs a continuation to
restore the scheduling group to its original value, but this
is redundant when executing our remote smp::submit_to() work,
since there is nothing chained to that continuation. We can
fire off that task in the correct scheduling group with
reactor::schedule() instead. Avoid allocating the task by
making the work_item a task.