tests: Delete Seastar.dist
This was disabled by test.py and it is currently failing. The original
intent was to test the packaged seastar, but to test that it would try
to download and compile a lot of code.

With the broken test gone it is now possible to run tests with just
"ninja test".

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200203163153.4503-1-espindola@scylladb.com>