reactor: remove kernel_completion::set_promise
As Avi suggested, we can rely on set_value alone.  There is a scenario
in which the I/O queue handler may want to set a failed promise before
dispatching the request at all. In which case it has its set_promise
just for itself.

Note that before this refactor we used to always call
io_desc::set_promise for the base class from the derived classes. As
we stopped doing that, we ate one call to _pr.set_exception in the
io_queue. That call is now restored.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200205151733.15499-1-glauber@scylladb.com>