net: Fix global buffer overflow around rss_key_type
The rsskey globals were defined as uint8_t[] but used as rss_key_type
(string_view). The default string_view constructor uses strlen, but
the keys are not null terminated.

This patch changes the global variables to be rss_key_type. With this
the string_view is constructed at compile time and uses the right
constructor.

Found while testing another patch with the native stack.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200205171125.429667-1-espindola@scylladb.com>