build: Use with -fstack-clash-protection when using guard pages
This option causes gcc to probe every page of a stack frame on
function entry by allocating one page at a time:

          leaq    -36864(%rsp), %r11
  .LPSRL0:
          subq    $4096, %rsp
          orq     $0, (%rsp)
          cmpq    %r11, %rsp
          jne     .LPSRL0
          subq    $2712, %rsp

This guarantees that one page of guard space is sufficient to catch a
stack overflow when it first occurs, which is of great help in
debugging such core files.

There are only a few functions that have stack frames larger than a
page, so this doesn't seem to have any impact on performance when
stack guards are enabled in release builds.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200206015508.770725-1-espindola@scylladb.com>