reactor: specialize the pollable_fd_state
A lot of the operations the reactor_backend needs to do take a
pollable_fd_state& as a parameter.

We would like to store data into the pollable_fd_state to aid with
completing those operations. We store the kernel_completion objects in
pollin and pollout attributes.

However as we add more capabilities to the backend, including the
upcoming uring implementation, it becomes clear that the backend may
need to store extra state in the pollable_fd_state that really is
backend-specific.

One example can be seen in this very patch: as we specialize the
pollable_fd_state per backend it is possible to store the iocb used for
the aio poll operations into the derived class itself, voiding the need
for us to dynamically allocate them.

When this change is made, one test actually fails: the alien test is
creating a pollable_fd_state directly, and operating on it. This
shouldn't be allowed, as pollable_fd_state is internal state for
pollable_fd. It also shouldn't be using read_some in an eventfd

To make sure this doesn't become a problem for anyone else, this patch
also makes the constructor for pollable_fd_state protected: now only the
backend specializations will be able to produce them.

Tests: unit (debug)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200206130106.11749-1-glauber@scylladb.com>