smp: introduce smp_submit_to_options
Intended to hold all options modifying the behaviour of
`smp::submit_to()`. For now it contains the service-group and the
timeout. Allows for seamless addition of new such options.

To allow a seamless transition from `smp_service_group` to
`smp_submit_to_options` in the existing `smp::submit_to()` and friend's
signatures, it is implicitly constructible from `smp_service_group`.