distributed_tests: increase sleep time further
Sleep-based tests are problematic. In commit 2398994aed55a0,
Avi increased the initial 10ms threshold to 30ms. However
during my rework of the reactor backend for uring, this start
started failing in debug mode.

It is tempting to think that I just made things slower, but:
- I don't see the slowdown in any benchmark
- It works for smp > 2, and only really fails for smp == 2 (because
  it has to execute more tasks)
- If I increase to as much as 33ms it works again
- The patch in the series in which it starts to fail really does
  nothing that could explain this.

So we were probably around the edge anyway and this pushed it.

I spent some time trying to see how we could make this less
dependent on time, but nothing worked well. The test is tricky
because we really want to see how much work we can push in parallel.
As tasks start to finish, the concurrency gets reduced. debug mode
reorders tasks, and anything can happen.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200206235002.31034-1-glauber@scylladb.com>