reactor: move lifetime management of the pollable_fd_state to backend
The reactor_backend is now creating the pollable_fd_state.
It should also be responsible for deleting it.

Doing that will allow us to issue asynchronous operations upon
destruction while keeping the object alive.

The candidate user is io_uring, which will want to issue an
asynchronous syscall to forget the pollable_fd

Signed-off-by: Glauber Costa <glauber@scylladb.com>