tests: stall_detector_test: use same clock as in test as in the detector
The stall detector uses the thread cputime clock to avoid reporting
stalls when the CPU was taken from the reactor by the scheduler to
run another process (those are also stalls, but reporting them as
the fault of some task is very confusing). But the test used to
verify the stall detector uses the realtime clock, so if the test
machine is overloaded (as is usual with continuous integration),
the stall detector fails to detect the stalls the test tries to
induce as not enough cputime has passed (even though enough realtime
has passed).

Fix by using the same clock in the test as the detector uses.

Tested by running 20 instances of the test in parallel on my
1s6c12t desktop. This fails consistently before the change and
passes afterwards. Hopefully this will fix the false positives
reported by CircleCI.
Message-Id: <20200209145041.647681-1-avi@scylladb.com>