http: Add the possibility to remove matches from routes
Currently in routes there are two types of rules -- exact
match and rule_match.

The former one is the direct match by url path and the handlers
sit in the unordered_map. For it -- add the .drop() method that
will just erase the entry from it (and return its former contents
to that the caller could free it).

The latter (rule match) is a bit more complex. There's no way
to directly map url to the entry, as they are looked-up not only
by URL, but by a list of substring matches. To remove this kind
of rule we use the introduced in the previous patch cookie. Two
methods are added -- the .add_cookie adds the match_rule object
and returns its cookie, the .del_cookie method accepts the cookie
and removes the entry matching it (also returning the old contents
to the caller).

The path_description can now have the .unset() method that will
unregister itself from the routes.

Also included are RAII-style wrappers for rules (both direct and
match) registration and auto-unregistration are included and unit
tests (using the mentioned wrappers).

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>