connect_test: keep socket alive until the end.
The socket assigned to unconn should be kept alive until the
end of the test. After shutdown, the statement "return f" causes
the test function to return, which can potentially destroy the
socket before the future has the opportunity to return.

This is not a problem today because connect() keeps the socket
shared pointer alive. This behavior may change, but even if it
doesn't it is still safer for the user to keep the pointer alive.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200211220053.30284-1-glauber@scylladb.com>