utils: add tmp_file
Users are advised to use tmp_file::do_with() to automatically
create the temporary file, and eventually close and remove it.

make_tmp_file can also be used to create a tmp_file,
yet the caller is responsible to close and remove it
in this case.

Temporary files are created by default in default_tmpdir().
This directory is determined by the TMPDIR environment variable,
if available, or `/tmp` otherwise.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>