move net::packet version of write_some close to the reactor methods
We didn't have a reactor call for write_some(net::packet&), and its
implementation was a method of pollable_fd_state. As we implement
that in the backend, we now want to keep the generic version in
the reactor, to make it easier for the backend to consume.

The previous patch kept everything in the same place to aid
reviewers. In this patch we move it closer to the other reactor
methods to aid new code readers.

Signed-off-by: Glauber Costa <glauber@scylladb.com>