reactor: don't call set_heap_profiling_enable() if not needed
Seastar has a `heapprof` configuration option, which will cause the
reactor to enable heap-profiling on all shards. It does this by
unconditionally calling `memory::set_heap_profiling_enable()` with the
value of the configuration option, which might be false.
Since a091b5d5 `memory::set_heap_profiling_enable()` will print a
warning if called when `SEASTAR_HEAPPROF` is not defined. This results
in this message being printed on startup from each shard, which is
confusing and unexpected when one didn't set the `heapprof`
configuration at all. Change the code to only call
`memory::set_heap_profiling_enable()` if `heapprof` was set to true,
avoiding the noise when no heap-profiling is attempted to be used.

Signed-off-by: Botond Dénes <bdenes@scylladb.com>
Message-Id: <20200212160918.1091986-1-bdenes@scylladb.com>