posix: change the signature of accept/try_accept
accept() and try_accept use addr_length, a member of the socket_address,
as a value-result argument. However the way this is coded today is that
we pass the length parameter is an arbitrary memory area and then copy
it back to addr_length.

We already moved away in a previous patch from calling try_accept like
this, as it is confusing.

This patch simplifies things further by changing the signature of accept
and try_accept so there is no room for misunderstandings.

There are no users of accept() in tree so that's not patched.

Thanks Rafael for the suggestion.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200212220314.18240-1-glauber@scylladb.com>