future: Don't call need_preempt in then/then_wrapped
This remove the last call to need_preempt in then and
then_wrapped. With this ready futures always call their continuation
directly.

This is an important API change. Before this change a long chain of
calls to future::then with ready futures would be interrupted if it
was taking too long or if there was a high priority task (like
reclaiming memory).

Now it is the responsibility of each user to call need_preempt in code
that might need too much time or memory before producing a non ready
future.

This is a 1.35% tps improvement in 500 randomized runs of

perf_simple_query --duration 16 --smp 1 -m4G

It is also a 3.4% reduction in the text size of scylla.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>