dns_test: Test with seastar.io instead of www.google.com
It has been observed that a DNS lookup of www.google.com can return an
answer containing 216.239.32.117 and that a reverse dns lookup on that
address fails.

In an attempt to make the dns test more stable, change it to use
seastar.io instead.

While at it, delete a bit of dead code.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200217221433.384300-1-espindola@scylladb.com>