future: Don't call need_preempt in promise_base::make_ready
We only need to make sure that need_preempt is called once per loop
iteration. Calling it on every continuation is a bit much. This patch
removes one of the calls to need_preempt reachable from
then/then_wrapped.

This is a 0.45% tps improvement in 500 randomized runs of

perf_simple_query --duration 16 --smp 1 -m4G

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>