core: add buffer_allocator interface
A buffer_allocator is used to decouple allocating buffers - from
where we allocate, and what size of buffer to allocate, from the
point of allocation. This allows us to allocate late (at the point
we know data is ready to be placed in the buffer) deep in the reactor
guts, but still use allocation strategies that the reactor does
not have control over (and may ultimately come from the user
program).

The buffer is added as an internal interface, but perhaps one day we
can expose it to applications.