posix_stack: do not allocate buffers for idle connections
Currently, posix_data_source_impl allocates a buffer and then waits
for the socket to become readable. This was forced by the reactor design
requiring read_some() to have a buffer supplied beforehand. But with
the read_some() overload taking a buffer_allocator, we can let the
reactor allocate the buffer after the socket has proven to be non-idle.

This can reduce memory consumption of servers that have many idle
connections.  In turn, we can later use this to increase buffer sizes
and so reduce syscall counts.