future: Always schedule in then/then_wrapped in debug builds
This brings back the old behavior where in a debug build then and
then_wrapped always schedule instead of executing the continuation in
place.

Doing this as a followup patch makes it easy to guarantee that no
change is done to non debug builds. In particular, it is still not
possible to schedule a ready future in a non debug build.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>