core: Add a on_internal_error helper
In scylla there are places where we want to check an invariant, but
not bring down the database if it fails.

Like the invariants checked with assert, these are things that should
never happen and indicate bugs, not just unusual issues like a
corrupted file.

Unlike the invariants checked with assert, we are reasonably confident
that it is possible to abort a particular task without bringing down
the entire server.

To handle that, scylla has a on_internal_error function that will
either throw or abort depending on a runtime flag.

This patch moves on_internal_error to seastar. It seems general enough
and we want this flag to be always set to abort during tests, which is
more naturally done in seastar.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200221213414.506150-2-espindola@scylladb.com>