Close connection, when reach to max retransmits
wait_send_available()  might wait for a future that will never be fulfilled. In case ,max_queue_space is exceeded and the peer is down (i.e. no acks), the future will never be fulfilled. Even when retransmitions are exhausted this future is not handled.
This patch replaces calling to cleanup() with calling to do_reset(), when re-transmitions are completed. do_reset() set an exception for this future.

Signed-off-by: Avner Taieb <avner@reduxio.com>
Acked-by: Asias He <asias@scylladb.com>
Message-Id: <20200223160811.34363-1-avner@reduxio.com>