httpd: Turn match_rule-s vector into map
The rule_match-es is currently a vector that's scanned
linearly till the first match, then the handler is returned.
The next patch needs to add the possibility to remove a rule
from it. For this the vector doesn't work well, as we can
only safely remove from it in O(n) time, which can drive us
into O(n^2) for one-by-one rules removal.

Also, the removal should be "re-entrable", i.e. gracefully
handle cases when two removals with the same "key" happen.

With the above requirements the new way of keeping the
sequence of rule_match-s is in the std::map, where the "key"
for the map is the constantly increasing 64bit integer
called cookie.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>