sstring: Add and use a uninitialized_string global function
This avoids using the initialized_later constructor directly. The
advantage being that uninitialized_string function still works with
std::string, it is just not as efficient since it returns an
initialized string.

There is an extension in libc++ that we could use for producing
uninitialized string. The extension is proposed for standard c++, so
we should be able to use it in libstdc++ too some time in the future.

I reported https://gcc.gnu.org/bugzilla/show_bug.cgi?id=94032 to track
the implementation in libstdc++.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>