http: Store a sstring in handler_registration
Storing a const reference to sstring is super dangerous since passing
a string literal causes a sstring to be constructed and immediately
destroyed.

Given the current uses this could be a std::string_view, but it seems
safer to just own a sstring.

This fixes a use-after-scope in httpd_test.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>