sstring: Pass string_view to compare
The real API of std::string is substantially more complicated as it
has to bolt support for string_view on top of the pre-c++11 API.

The sstring implementation in seastar doesn't have all the overloads
of compare, and it probably should not.

This patch then just replaces a const sstring reference with a
compat::basic_string_view.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200311004021.269968-5-espindola@scylladb.com>