reactor: always initialize smp_poller
alien polling functions are called from within the smp poller, see
reactor::smp_pollfn. if a single core is designated to reactor,
it's fine as long as we do not use alien::submit_to(), as no cross-core
communication is needed. but if alien::submit_to() is used, we do
need the smp poller even if smp::count is 1. as we hav to check the
task submitted by an alien thread where the reactor is not available.

so in this change, smp_poller is always created and registered in
reactor even if smp::count is 1.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>
Message-Id: <20200311035642.15084-2-tchaikov@gmail.com>