reactor: Use do_with to save stack space
The "retries" variable is a static_vector, which is a very large
type. This was wastefull since we were copying to the stack just to
produce two more copies, one in each lambda.

As far as I can tell it is valid to use the same buffer for both
lambdas, so this patch uses a do_with to produce just one copy into a
heap allocated vector and the lambdas have references to it.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200311220741.363569-4-espindola@scylladb.com>