future: Add an explicit promise member to continuation
Before we were storing the promise inside the lambda. Storing it
outside should cost us nothing more and might even be a small will in
reduced code duplication as the promise is now in a base class.

The advantage of making the promise explicit is that it can now be
accessed more easily, which will be useful for traversing from a
continuation to the promise that the continuation will make available
when executed.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200317220054.2039147-1-espindola@scylladb.com>