smp: optimize this_shard_id()
this_shard_id() is simple enough to be compiled into a single
instruction, so optimize it so.

To avoid problems with the compiler generating initialization
guards, we place the thread_local variable in the same header, so
the compiler can see it does not require dynamic initialization.

The motivation for optimizing this_shard_id() is that if we use
it to replace engine().cpu_id() calls, most inclusions of reactor.hh
will disappear.