scheduling_group_specific: remove inclusion of reactor.hh
Since reactor.hh is heavyweight and full of implementation
details, including it is detrimental. This patch removes such
an include from scheduling_specific.hh. It is complex since
we have to mirror the per-task-queue structure without resorting
to referencing task_queues, as they are part of reactor.hh.

To avoid initialization order time problems, the data is still
hosted by class reactor; it exposes a pointer to it through a
thread_local variable.

Tests: unit (dev, debug)
Message-Id: <20200322173408.754165-1-avi@scylladb.com>