semaphore_units: allow getting the number of units held
A wrapper for held semaphore units - semaphore_units class
can now be queried for the number of units it holds,
which can be useful e.g. in order to transfer these units
to another semaphore, possibly on another shard.
This commit comes with a simple test case.

Tests: unit(dev)
Message-Id: <9d893b0d4f4943342ae4eebc886bc38e839d2091.1585035615.git.sarna@scylladb.com>