io_queue: fix miscalculation of sizes when I/O queue is not configured.
Although this is technically an unsupported scenario (the I/O queue must
be configured to either restrict by capacity or by iops + bandwidth),
the current code as it is makes it so that an unconfigured I/O queue
always set write requests to have 0 as their size and weight.

That's because in the absence of a configured capacity, we take the
first branch. In there, we multiply the write bandwidth/iops by a
constant factor and then divide it by the read bandwidth/iops.

In a misconfigured io_queue, the measured quantities will be
uint64_t::max() and the first multiplication will overflow, yielding
zero as a result.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200326201012.7871-1-glauber@scylladb.com>