file: use correct io_queue in dup()d files
We are dispatching requests through the wrong I/O queue if the
file is created from a file_handle, such as when dup() is called
on, and used in a new shard.

This is because we preserve I/O queue information in the file_handle.
But different shards will have different I/O queues.

This patch removes the io_queue pointer from the file_handle altogether,
replacing it with the information about the device id. The io_queue is
then always created from the device id in the posix_file_impl itself.

Fixes #730

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200327171236.16875-1-glauber@scylladb.com>