json: Add backtrace information for json generation exceptions
The json generation infrasructure in seastar, uses to_json
overloaded methods to do part of it's generation magic.
Those function are not prototyped as noexcept so it is possible
and indeed happens that while generating a json object, one of
the overloads will throw.
A concrete example is the to_json overloaded function for the double
type which cannot handle infinite values.
When those exceptions do happen we would like to have a some lead
about which field caused the error so we can take measures to
fix the values or rethink the specific generation.

This commit adds information by nesting an exception
with the field name that caused the to_json function to throw.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>
Message-Id: <20200329161021.242853-1-eliransin@scylladb.com>