reactor: decouple idle cpu handler from reactor
Provide a reactor-free entry point and supporting types for the
idle cpu handler. This allows users to invoke it without including
reactor.hh, which is quite heavy.

The supporting types are the more important parts of this path, as
they will often be mentioned in headers, while the API entry point
is usually needed from just one translation unit.

As a bonus, document the supporting types.
Message-Id: <20200329172949.13135-1-avi@scylladb.com>