file: Keep file alive until close completes
This replaces the traditional ".finally([f] { })" with a centralized
fix.

We only have to handle the close member function as a file must always
be closed, by which point the futures returned by other member
functions will have completed too.

In addition to the easy of use, I think a rationale for doing this is:

Lifetime management in c++ is complicated, and adding futures makes it
even more so. There is no one single rule that if followed solves it,
but there are a few useful guidelines.

Probably the first thing one should try is unique ownership. With that
the owner is responsible for keeping a value alive until all futures
using the value complete.

The next option is shared ownership, which is what we use for
files. In here the question is what should be responsible for
guaranteeing that a reference is kept until a future resolves.

One option, which we have now, is to make it as similar to unique
ownership as possible. This pushes the responsibility out in the call
stack, so that a function that takes file by value and schedules
multiple continuations, is responsible for keeping one reference alive
until all continuations it scheduled finish.

Another option, which this patch implements, is to say that since we
are using shared ownership anyway, the continuation closing a file
should take part in the reference count keeping the file alive.

This seems more in line with the spirit of shared ownership and, as
far as I can tell, similar to what is done by systems that manage
reference counts automatically
(https://en.wikipedia.org/wiki/Automatic_Reference_Counting).

Fixes #600

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>