io_queue: batch communication with the fair_queue for ready requests
Every time a request is ready, and right after it is ready, we tell the
underlying fair_queue that the request is ready. There is nothing wrong
with that, because this is a quite cheap operation.

This is soon going to become more expensive as I move towards supporting
shared and nested queues. In preparation for that, I am moving the
communication between the local I/O queue and the shared queue to a single
point at the end of the poller.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200403182107.30401-1-glauber@scylladb.com>