fair_queue rename the fair_queue_descriptor and change its default init
The fair_queue_request_descriptor can now represent either a request or
a group of requests. The name fair_queue_ticket is then more
appropriate. With that, there is no reason to believe that "1" is a
better default than anything else and we then start zero-initializing
this structure.

The signature of notify_requests_finished() is changed so we take
an object, not a reference. This allow us to write:

  _fq.notify_requests_finished(std::exchange(_completed_accumulator, {}));

in the io_queue code. With an lvalue signature, we'd have to
keep an object around to pass to notify_requests_finished().

Tests: unit (dev)

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20200406161853.16278-1-glauber@scylladb.com>